Rename useList to useCheckList
